QStorageInfo/Unix: fix narrowing conversion warning
"blocksize" is assigned the value of statvfs::f_bsize, the latter is
unsigned long.

+++ src/corelib/io/qstorageinfo_p.h
@@ -58,7 +58,7 @@ public:
     qint64 bytesTotal;
     qint64 bytesFree;
     qint64 bytesAvailable;
-    int blockSize;
+    ulong blockSize;
 
     bool readOnly;
     bool ready;
